Extract generic map internals into new gmap module
